moves merger formatting function back into merger #379
